requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
Django==5.2.10
django-debug-toolbar==6.2.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
greenlet==3.3.0
playwright==1.57.0
psycopg2-binary==2.9.11
pyee==13.0.0
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
